+++ js/services/air-quality.service.js
@@ -1,4 +1,4 @@
-import { APP_CONFIG } from "../../config/config.js?v=1.5.0-release";
+import { APP_CONFIG } from "../../config/config.js?v=1.5.1-release";
 
 const AIR_QUALITY_VARIABLES = [
     "european_aqi",